Moved pages into src/pages.
